Don't clamp when creating condition functions
There are situations where the value in a conditional expression lies
outwith the numerical range for the attribute against which it is to be
applied. A clamp was therefore applied when creating the condition
function. Unfortunately performing this clamp is not necessarily correct
when the numerical range of the attribute subsequently changes. So
instead we just avoid the clamp altogether here. The value is clamped at
other points in the transform system, so it's not really a danger to
remove its application here.