Fix crash in GraphComponentRenderer::moveFocusToNode
In switching to component mode via...

GraphRenderer::synchronize
GraphRenderer::moveFocusToNode
GraphComponentRenderer::moveFocusToNode

...it was possible for GraphComponentRenderer::_nodeIds to not be
up-to-date with respect to the graph, because the call to ::synchronize
preceeds the renderer's deferred execution pipeline, in which _nodeIds
was updated. Instead we now focus the node via
GraphComponentRenderer::switchToComponentMode, which *is* executed in
said deferred execution pipeline.