Call Graph::update after a clone
When doing a sequence of transforms, the graph is cloned for each
iteration, leaving it in an _updateRequired state. This means that when
it comes to actually applying the transform, an update is always
performed, and a change in state is always detected, even when nothing
actually changed; so now we unconditionally do an update after each
clone and changes are correctly detected.