Rearrange ThreadPool wait loop to avoid unlock/relock gap
This is just a stab in the dark really, to see if this fixes a
persistent crash on macOS in _waitForNewTask::wait. The previous
behaviour was to be woken up with the lock, then immediately release it
before reacquiring it on the next iteration. As far as I can reason,
this gap wouldn't have actually been harmful, but at the very least it
wasn't particularly neat.