Fix crash on debug builds under MSVC
MSVC has various extra tests to find STL misuse, one of which is the
usage of invalid iterators (_ITERATOR_DEBUG_LEVEL, on by default). When
computing the BH tree for a graph, the vector containing recently
processed trees is cleared and rebuilt. Simultaneously the ThreadPool's
packaged_task destructors are potentially running, one of whose actions
is to destruct the pairs of iterators each thread ranged over. The
destructors for said iterators perform some checks and a crash occurs
since the iterators are in the process of being invalidated on the main
thread. The crash is avoided here by resetting the iterators to default
immediately before they're used, so that when they come to be destructed
they can't interfere with anything happening on the main thread. It
seems a bit odd that destructing invalid iterators are able to crash so
this is potentially a bug in MS STL. However on the other hand
simultaneously doing anything with iterators from multiple threads is
most likely UB in the first place, so this change is probably a good one
regardless.