Disallow using non-user defined attributes as keys when importing
This fixes a crash that occurred when auto detecting key matches that
occurred because calculated attributes (usually) consult the transformed
graph for their values, but the element ids for the matches are sourced
from the base graph. If the latter is not in the former, this obviously
causes a problem.