Fix TransformParameter slider being apparently inoperative where the range is large
QQC1 Sliders have a strange bug where if its minimum or maximum is set
outside +/-21474836 (curiously 1% of the min/max of a 32 bit signed
integer), they still partially work when clicking and dragging, but the
button remains fixed to the left. To work around this, the Slider is now
given 0.0 and 1.0 for its minimum and maximum, and its value is lerped
and normalised as appropriate.