Fix PreferencesWatcher
PreferencesWatcher works by having a static list of watchers that are
notified whenever a change occurs. When part of the shared library this
means that the static lists get overwritten by the instances in the
library. This change moves PreferencesWatcher to live app side and
changes the shared utils preferences interface to only allow read
access.