Fix data race in ForceDirectedLayout
When calculating attractive forces between nodes connected by an edge,
we assigned a subset of the edges to each thread which then applied the
results to the nodes. If two distinct sets of edges referred to the same
node, this caused a data race where two threads would simultaneously
write to the attractive force result for the same node. To fix it, we
now store the results in an intermediate array, and apply them to the
nodes separately, immediately before the consolidation and damping step.