Remove -fvisibility=hidden from Unix compiler flags
Having this in place was causing a bunch of warnings on macOS when
dynamic_casting some of the types contained in the shared library. It's not
clear why it was added in the first place since the ODR violations it claims to
avoid presumably only occur in the static linking case, and the shared library
is not statically linked. Hmm.