Remove workaround for checked menu item state
Previously when using ButtonGroups and bound state for the menu items'
checked property, it would sometimes get into a situation where despite
the actual returned value of the state being correct, the visual
appearance of the menu item(s) would be wrong, so in order to mitigate
this a simple toggle was done on the item when the state needed to
change. Testing this again, it no longer seems to be necessary, but the
reasons this is so are not clear, hence this explanatory text, in case
it becomes a problem again in future.