Delete stale instances of AvailableAttributesModel
Various places that allocated instances of AvailableAttributesModel held
on to the pointers after they were needed, and if the attribute
landscape changed in the mean time, this could lead to crashes when
attribute pointers were accessed after they had been destroyed. Indeed,
this is probably the cause of a relatively common crash in
AvailableAttributesModel::data, the most likely instance of which
originating from the EnrichmentWizard.