Relax the correlation parser gatekeeping
When parsing a tabular data file, normally the column counts for the
first few lines are checked to see if they match, within some range. For
pairwise and matrix formats, this makes sense as empty cells in that
case indicate something is fairly wrong. For correlation on the other
hand it does not, particularly in the discrete case. The correlation
parser was inadvertently performing this check anyway, so now it is
disabled.