Ignore "numRows" in the save file
Turns out that if you manually delete nodes from a graph then save it,
the numRows value will always be larger than reality, since it doesn't
account for the deleted rows. This is fine in the live case since the
mapping between nodes and rows takes care of things, but when restoring
from a save file a 1:1 mapping is assumed, and numRows will always be
too large, so now we just ignore the saved value and infer what it
should be from the number of columns and the number of data points.