Fix crash when deleting nodes
Signals emitted by SelectionManager during node deselection could result
in a QML expression to be re-evaluated that in turn would query the
graph for a component's node count. If the timing was wrong this would
occur when component management was disabled, causing a crash. We avoid
this now by simply preventing the initial signal.