Improve Windows installer admin/user behaviour
If an installation has been performed as admin, but updates are
performed as a user then we can get into the situation where the
installer thinks there is no previous installation present, and the
update results in a second user copy being installer. The change here is
to explicitly detect an admin installation and use this context if there
is. Potentially this means a failed installation, but this is probably a
better UX than installing two copies.