Revert "Prevent crashes when TreeBox has invisible rows"
This reverts commit 11bddac7aaf0671b0e1c1539a72892d3fd21230b.

I have no recollection of the circumstances under which crashes
occurred, suffice it to say it appears as though the source of the
crashes is in TreeView.positionViewAtRow; hopefully any bugs here have
been addressed upstream.

The reason for reverting the commit is that it was preventing keyboard
navigation from working when searching a TreeBox. Pressing up or down
arrows would not refocus the TreeView onto the selected row.