Use a TBO on desktop and plain texture on web (workaround suspected NVidia driver bug)
On Windows desktop, using an NVidia video card, using a plain old
texture to buffer the component data results in a crash when calling
glTexImage2D from ::uploadGPUComponentData. It's not at all clear why
this is problematic, suffice it to say the driver is suspected to be
buggy. Instead therefore we revert to the previous method of using a TBO
on the desktop, only resorting to the plain old texture on web, where
TBOs aren't a thing.